First Posting of  Shuffle Array
This code shuffles large arrays for machine learning.
The goal of this code is to shuffle the arrays without
loading the entire array into to memory. Normally I would
use sklearn's shuffle as it is faster but it loads the entire array
into memory from what I can tell as it give me memory errors
on larger arrays on systems with smaller memory.